Move the impls for PollError into error.rs
These should have been moved when error.rs was created but were missed.
